kernel/generic: disable LZMA and RTIME in jffs2
LZMA itself weighs over 200 kB in a typical sysupgrade image. It's implemented
by a huge out-of-tree patch which has been forward-ported since the dawn of
OpenWrt, and has never been upstreamed. As for RTIME, it doesn't really buy us
anything, as we're not trying to compress already compressed data.

Disable both algorithms.

Signed-off-by: Rui Salvaterra <rsalvaterra@gmail.com>